refactor cypress into pub/sub instrumentation
